Moved building name generation into utils.
