+++ OpenTESArena/src/World/CityLevelUtils.cpp
@@ -1,9 +1,14 @@
 #include "CityLevelUtils.h"
 #include "LocationUtils.h"
+#include "VoxelDataType.h"
+#include "VoxelDefinition.h"
+#include "VoxelGrid.h"
 #include "../Assets/MIFFile.h"
 #include "../Assets/MIFUtils.h"
 #include "../Assets/MiscAssets.h"
 #include "../Math/Random.h"
+
+#include "components/utilities/String.h"
 
 void CityLevelUtils::generateCity(uint32_t citySeed, int cityDim, WEInt gridDepth,
 	const std::vector<uint8_t> &reservedBlocks, const OriginalInt2 &startPosition,
@@ -113,6 +118,221 @@ void CityLevelUtils::generateCity(uint32_t citySeed, int cityDim, WEInt gridDept
 			zDim++;
 		}
 	}
+}
+
+LevelUtils::MenuNamesList CityLevelUtils::generateBuildingNames(const LocationDefinition &locationDef,
+	const ProvinceDefinition &provinceDef, ArenaRandom &random, bool isCity,
+	const VoxelGrid &voxelGrid, const MiscAssets &miscAssets)
+{
+	const auto &exeData = miscAssets.getExeData();
+	const LocationDefinition::CityDefinition &cityDef = locationDef.getCityDefinition();
+
+	uint32_t citySeed = cityDef.citySeed;
+	const Int2 localCityPoint = LocationUtils::getLocalCityPoint(citySeed);
+
+	LevelUtils::MenuNamesList menuNames;
+
+	// Lambda for looping through main-floor voxels and generating names for *MENU blocks that
+	// match the given menu type.
+	auto generateNames = [&provinceDef, &citySeed, &random, isCity, &voxelGrid, &miscAssets,
+		&exeData, &cityDef, &localCityPoint, &menuNames](VoxelDefinition::WallData::MenuType menuType)
+	{
+		if ((menuType == VoxelDefinition::WallData::MenuType::Equipment) ||
+			(menuType == VoxelDefinition::WallData::MenuType::Temple))
+		{
+			citySeed = (localCityPoint.x << 16) + localCityPoint.y;
+			random.srand(citySeed);
+		}
+
+		std::vector<int> seen;
+		auto hashInSeen = [&seen](int hash)
+		{
+			return std::find(seen.begin(), seen.end(), hash) != seen.end();
+		};
+
+		// Lambdas for creating tavern, equipment store, and temple building names.
+		auto createTavernName = [&exeData, &cityDef](int m, int n)
+		{
+			const auto &tavernPrefixes = exeData.cityGen.tavernPrefixes;
+			const auto &tavernSuffixes = cityDef.coastal ?
+				exeData.cityGen.tavernMarineSuffixes : exeData.cityGen.tavernSuffixes;
+			return tavernPrefixes.at(m) + ' ' + tavernSuffixes.at(n);
+		};
+
+		auto createEquipmentName = [&provinceDef, &random, &miscAssets, &exeData,
+			&cityDef](int m, int n, SNInt x, WEInt z)
+		{
+			const auto &equipmentPrefixes = exeData.cityGen.equipmentPrefixes;
+			const auto &equipmentSuffixes = exeData.cityGen.equipmentSuffixes;
+
+			// Equipment store names can have variables in them.
+			std::string str = equipmentPrefixes.at(m) + ' ' + equipmentSuffixes.at(n);
+
+			// Replace %ct with city type name.
+			size_t index = str.find("%ct");
+			if (index != std::string::npos)
+			{
+				const std::string_view cityTypeName = cityDef.typeDisplayName;
+				str.replace(index, 3, cityTypeName);
+			}
+
+			// Replace %ef with generated male first name from (y<<16)+x seed. Use a local RNG for
+			// modifications to building names. Swap and reverse the XZ dimensions so they fit the
+			// original XY values in Arena.
+			index = str.find("%ef");
+			if (index != std::string::npos)
+			{
+				ArenaRandom nameRandom((x << 16) + z);
+				const bool isMale = true;
+				const std::string maleFirstName = [&provinceDef, &miscAssets, isMale, &nameRandom]()
+				{
+					const std::string name = miscAssets.generateNpcName(
+						provinceDef.getRaceID(), isMale, nameRandom);
+					const std::string firstName = String::split(name).front();
+					return firstName;
+				}();
+
+				str.replace(index, 3, maleFirstName);
+			}
+
+			// Replace %n with generated male name from (x<<16)+y seed.
+			index = str.find("%n");
+			if (index != std::string::npos)
+			{
+				ArenaRandom nameRandom((z << 16) + x);
+				const bool isMale = true;
+				const std::string maleName = miscAssets.generateNpcName(
+					provinceDef.getRaceID(), isMale, nameRandom);
+				str.replace(index, 2, maleName);
+			}
+
+			return str;
+		};
+
+		auto createTempleName = [&exeData](int model, int n)
+		{
+			const auto &templePrefixes = exeData.cityGen.templePrefixes;
+			const auto &temple1Suffixes = exeData.cityGen.temple1Suffixes;
+			const auto &temple2Suffixes = exeData.cityGen.temple2Suffixes;
+			const auto &temple3Suffixes = exeData.cityGen.temple3Suffixes;
+
+			const std::string &templeSuffix = [&temple1Suffixes, &temple2Suffixes,
+				&temple3Suffixes, model, n]() -> const std::string&
+			{
+				if (model == 0)
+				{
+					return temple1Suffixes.at(n);
+				}
+				else if (model == 1)
+				{
+					return temple2Suffixes.at(n);
+				}
+				else
+				{
+					return temple3Suffixes.at(n);
+				}
+			}();
+
+			// No extra whitespace needed, I think?
+			return templePrefixes.at(model) + templeSuffix;
+		};
+
+		// The lambda called for each main-floor voxel in the area.
+		auto tryGenerateBlockName = [menuType, &random, isCity, &voxelGrid, &menuNames, &seen,
+			&hashInSeen, &createTavernName, &createEquipmentName, &createTempleName](SNInt x, WEInt z)
+		{
+			// See if the current voxel is a *MENU block and matches the target menu type.
+			const bool matchesTargetType = [x, z, menuType, isCity, &voxelGrid]()
+			{
+				const uint16_t voxelID = voxelGrid.getVoxel(x, 1, z);
+				const VoxelDefinition &voxelDef = voxelGrid.getVoxelDef(voxelID);
+				return (voxelDef.dataType == VoxelDataType::Wall) && voxelDef.wall.isMenu() &&
+					(VoxelDefinition::WallData::getMenuType(voxelDef.wall.menuID, isCity) == menuType);
+			}();
+
+			if (matchesTargetType)
+			{
+				// Get the *MENU block's display name.
+				int hash;
+				std::string name;
+
+				if (menuType == VoxelDefinition::WallData::MenuType::Tavern)
+				{
+					// Tavern.
+					int m, n;
+					do
+					{
+						m = random.next() % 23;
+						n = random.next() % 23;
+						hash = (m << 8) + n;
+					} while (hashInSeen(hash));
+
+					name = createTavernName(m, n);
+				}
+				else if (menuType == VoxelDefinition::WallData::MenuType::Equipment)
+				{
+					// Equipment store.
+					int m, n;
+					do
+					{
+						m = random.next() % 20;
+						n = random.next() % 10;
+						hash = (m << 8) + n;
+					} while (hashInSeen(hash));
+
+					name = createEquipmentName(m, n, x, z);
+				}
+				else
+				{
+					// Temple.
+					int model, n;
+					do
+					{
+						model = random.next() % 3;
+						const std::array<int, 3> ModelVars = { 5, 9, 10 };
+						const int vars = ModelVars.at(model);
+						n = random.next() % vars;
+						hash = (model << 8) + n;
+					} while (hashInSeen(hash));
+
+					name = createTempleName(model, n);
+				}
+
+				menuNames.push_back(std::make_pair(NewInt2(x, z), std::move(name)));
+				seen.push_back(hash);
+			}
+		};
+
+		// Start at the top-right corner of the map, running right to left and top to bottom.
+		for (SNInt x = 0; x < voxelGrid.getWidth(); x++)
+		{
+			for (WEInt z = 0; z < voxelGrid.getDepth(); z++)
+			{
+				tryGenerateBlockName(x, z);
+			}
+		}
+
+		// Fix some edge cases used with the main quest.
+		if ((menuType == VoxelDefinition::WallData::MenuType::Temple) &&
+			cityDef.hasMainQuestTempleOverride)
+		{
+			const auto &mainQuestTempleOverride = cityDef.mainQuestTempleOverride;
+			const int modelIndex = mainQuestTempleOverride.modelIndex;
+			const int suffixIndex = mainQuestTempleOverride.suffixIndex;
+
+			// Added an index variable since the original game seems to store its menu names in a
+			// way other than with a vector like this solution is using.
+			const int menuNamesIndex = mainQuestTempleOverride.menuNamesIndex;
+
+			DebugAssertIndex(menuNames, menuNamesIndex);
+			menuNames[menuNamesIndex].second = createTempleName(modelIndex, suffixIndex);
+		}
+	};
+
+	generateNames(VoxelDefinition::WallData::MenuType::Tavern);
+	generateNames(VoxelDefinition::WallData::MenuType::Equipment);
+	generateNames(VoxelDefinition::WallData::MenuType::Temple);
+	return menuNames;
 }
 
 void CityLevelUtils::revisePalaceGraphics(Buffer2D<uint16_t> &map1, SNInt gridWidth, WEInt gridDepth)
