+++ OpenTESArena/src/World/ExteriorLevelData.cpp
@@ -29,341 +29,6 @@ ExteriorLevelData::~ExteriorLevelData()
 
 }
 
-void ExteriorLevelData::generateBuildingNames(const LocationDefinition &locationDef,
-	const ProvinceDefinition &provinceDef, ArenaRandom &random, bool isCity,
-	SNInt gridWidth, WEInt gridDepth, const MiscAssets &miscAssets)
-{
-	const auto &exeData = miscAssets.getExeData();
-	const LocationDefinition::CityDefinition &cityDef = locationDef.getCityDefinition();
-	
-	uint32_t citySeed = cityDef.citySeed;
-	const Int2 localCityPoint = LocationUtils::getLocalCityPoint(citySeed);
-
-	// Lambda for looping through main-floor voxels and generating names for *MENU blocks that
-	// match the given menu type.
-	auto generateNames = [this, &provinceDef, &citySeed, &random, isCity,
-		gridWidth, gridDepth, &miscAssets, &exeData, &cityDef, &localCityPoint](
-			VoxelDefinition::WallData::MenuType menuType)
-	{
-		if ((menuType == VoxelDefinition::WallData::MenuType::Equipment) ||
-			(menuType == VoxelDefinition::WallData::MenuType::Temple))
-		{
-			citySeed = (localCityPoint.x << 16) + localCityPoint.y;
-			random.srand(citySeed);
-		}
-
-		std::vector<int> seen;
-		auto hashInSeen = [&seen](int hash)
-		{
-			return std::find(seen.begin(), seen.end(), hash) != seen.end();
-		};
-
-		// Lambdas for creating tavern, equipment store, and temple building names.
-		auto createTavernName = [&exeData, &cityDef](int m, int n)
-		{
-			const auto &tavernPrefixes = exeData.cityGen.tavernPrefixes;
-			const auto &tavernSuffixes = cityDef.coastal ?
-				exeData.cityGen.tavernMarineSuffixes : exeData.cityGen.tavernSuffixes;
-			return tavernPrefixes.at(m) + ' ' + tavernSuffixes.at(n);
-		};
-
-		auto createEquipmentName = [&provinceDef, &random, gridWidth, gridDepth, &miscAssets,
-			&exeData, &cityDef](int m, int n, SNInt x, WEInt z)
-		{
-			const auto &equipmentPrefixes = exeData.cityGen.equipmentPrefixes;
-			const auto &equipmentSuffixes = exeData.cityGen.equipmentSuffixes;
-
-			// Equipment store names can have variables in them.
-			std::string str = equipmentPrefixes.at(m) + ' ' + equipmentSuffixes.at(n);
-
-			// Replace %ct with city type name.
-			size_t index = str.find("%ct");
-			if (index != std::string::npos)
-			{
-				const std::string_view cityTypeName = cityDef.typeDisplayName;
-				str.replace(index, 3, cityTypeName);
-			}
-
-			// Replace %ef with generated male first name from (y<<16)+x seed. Use a local RNG for
-			// modifications to building names. Swap and reverse the XZ dimensions so they fit the
-			// original XY values in Arena.
-			index = str.find("%ef");
-			if (index != std::string::npos)
-			{
-				ArenaRandom nameRandom((x << 16) + z);
-				const bool isMale = true;
-				const std::string maleFirstName = [&provinceDef, &miscAssets, isMale, &nameRandom]()
-				{
-					const std::string name = miscAssets.generateNpcName(
-						provinceDef.getRaceID(), isMale, nameRandom);
-					const std::string firstName = String::split(name).front();
-					return firstName;
-				}();
-
-				str.replace(index, 3, maleFirstName);
-			}
-
-			// Replace %n with generated male name from (x<<16)+y seed.
-			index = str.find("%n");
-			if (index != std::string::npos)
-			{
-				ArenaRandom nameRandom((z << 16) + x);
-				const bool isMale = true;
-				const std::string maleName = miscAssets.generateNpcName(
-					provinceDef.getRaceID(), isMale, nameRandom);
-				str.replace(index, 2, maleName);
-			}
-
-			return str;
-		};
-
-		auto createTempleName = [&exeData](int model, int n)
-		{
-			const auto &templePrefixes = exeData.cityGen.templePrefixes;
-			const auto &temple1Suffixes = exeData.cityGen.temple1Suffixes;
-			const auto &temple2Suffixes = exeData.cityGen.temple2Suffixes;
-			const auto &temple3Suffixes = exeData.cityGen.temple3Suffixes;
-
-			const std::string &templeSuffix = [&temple1Suffixes, &temple2Suffixes,
-				&temple3Suffixes, model, n]() -> const std::string&
-			{
-				if (model == 0)
-				{
-					return temple1Suffixes.at(n);
-				}
-				else if (model == 1)
-				{
-					return temple2Suffixes.at(n);
-				}
-				else
-				{
-					return temple3Suffixes.at(n);
-				}
-			}();
-
-			// No extra whitespace needed, I think?
-			return templePrefixes.at(model) + templeSuffix;
-		};
-
-		// The lambda called for each main-floor voxel in the area.
-		auto tryGenerateBlockName = [this, isCity, menuType, &random, &seen, &hashInSeen,
-			&createTavernName, &createEquipmentName, &createTempleName](SNInt x, WEInt z)
-		{
-			// See if the current voxel is a *MENU block and matches the target menu type.
-			const bool matchesTargetType = [this, isCity, x, z, menuType]()
-			{
-				const auto &voxelGrid = this->getVoxelGrid();
-				const uint16_t voxelID = voxelGrid.getVoxel(x, 1, z);
-				const VoxelDefinition &voxelDef = voxelGrid.getVoxelDef(voxelID);
-				return (voxelDef.dataType == VoxelDataType::Wall) && voxelDef.wall.isMenu() &&
-					(VoxelDefinition::WallData::getMenuType(voxelDef.wall.menuID, isCity) == menuType);
-			}();
-
-			if (matchesTargetType)
-			{
-				// Get the *MENU block's display name.
-				int hash;
-				std::string name;
-
-				if (menuType == VoxelDefinition::WallData::MenuType::Tavern)
-				{
-					// Tavern.
-					int m, n;
-					do
-					{
-						m = random.next() % 23;
-						n = random.next() % 23;
-						hash = (m << 8) + n;
-					} while (hashInSeen(hash));
-
-					name = createTavernName(m, n);
-				}
-				else if (menuType == VoxelDefinition::WallData::MenuType::Equipment)
-				{
-					// Equipment store.
-					int m, n;
-					do
-					{
-						m = random.next() % 20;
-						n = random.next() % 10;
-						hash = (m << 8) + n;
-					} while (hashInSeen(hash));
-
-					name = createEquipmentName(m, n, x, z);
-				}
-				else
-				{
-					// Temple.
-					int model, n;
-					do
-					{
-						model = random.next() % 3;
-						const std::array<int, 3> ModelVars = { 5, 9, 10 };
-						const int vars = ModelVars.at(model);
-						n = random.next() % vars;
-						hash = (model << 8) + n;
-					} while (hashInSeen(hash));
-
-					name = createTempleName(model, n);
-				}
-
-				this->menuNames.push_back(std::make_pair(NewInt2(x, z), std::move(name)));
-				seen.push_back(hash);
-			}
-		};
-
-		// Start at the top-right corner of the map, running right to left and top to bottom.
-		for (SNInt x = 0; x < gridWidth; x++)
-		{
-			for (WEInt z = 0; z < gridDepth; z++)
-			{
-				tryGenerateBlockName(x, z);
-			}
-		}
-
-		// Fix some edge cases used with the main quest.
-		if ((menuType == VoxelDefinition::WallData::MenuType::Temple) &&
-			cityDef.hasMainQuestTempleOverride)
-		{
-			const auto &mainQuestTempleOverride = cityDef.mainQuestTempleOverride;
-			const int modelIndex = mainQuestTempleOverride.modelIndex;
-			const int suffixIndex = mainQuestTempleOverride.suffixIndex;
-
-			// Added an index variable since the original game seems to store its menu names in a
-			// way other than with a vector like this solution is using.
-			const int menuNamesIndex = mainQuestTempleOverride.menuNamesIndex;
-
-			DebugAssertIndex(this->menuNames, menuNamesIndex);
-			this->menuNames[menuNamesIndex].second = createTempleName(modelIndex, suffixIndex);
-		}
-	};
-
-	generateNames(VoxelDefinition::WallData::MenuType::Tavern);
-	generateNames(VoxelDefinition::WallData::MenuType::Equipment);
-	generateNames(VoxelDefinition::WallData::MenuType::Temple);
-}
-
-void ExteriorLevelData::generateWildChunkBuildingNames(const ExeData &exeData)
-{
-	// Lambda for looping through main-floor voxels and generating names for *MENU blocks that
-	// match the given menu type.
-	auto generateNames = [this, &exeData](int wildX, int wildY,
-		VoxelDefinition::WallData::MenuType menuType)
-	{
-		const uint32_t wildChunkSeed = (wildY << 16) + wildX;
-
-		// Don't need hashInSeen() for the wilderness.
-
-		// Lambdas for creating tavern and temple building names.
-		auto createTavernName = [&exeData](int m, int n)
-		{
-			const auto &tavernPrefixes = exeData.cityGen.tavernPrefixes;
-			const auto &tavernSuffixes = exeData.cityGen.tavernSuffixes;
-			return tavernPrefixes.at(m) + ' ' + tavernSuffixes.at(n);
-		};
-
-		auto createTempleName = [&exeData](int model, int n)
-		{
-			const auto &templePrefixes = exeData.cityGen.templePrefixes;
-			const auto &temple1Suffixes = exeData.cityGen.temple1Suffixes;
-			const auto &temple2Suffixes = exeData.cityGen.temple2Suffixes;
-			const auto &temple3Suffixes = exeData.cityGen.temple3Suffixes;
-
-			const std::string &templeSuffix = [&temple1Suffixes, &temple2Suffixes,
-				&temple3Suffixes, model, n]() -> const std::string&
-			{
-				if (model == 0)
-				{
-					return temple1Suffixes.at(n);
-				}
-				else if (model == 1)
-				{
-					return temple2Suffixes.at(n);
-				}
-				else
-				{
-					return temple3Suffixes.at(n);
-				}
-			}();
-
-			// No extra whitespace needed, I think?
-			return templePrefixes.at(model) + templeSuffix;
-		};
-
-		// The lambda called for each main-floor voxel in the area.
-		auto tryGenerateBlockName = [this, wildX, wildY, wildChunkSeed, menuType, &createTavernName,
-			&createTempleName](SNInt x, WEInt z)
-		{
-			ArenaRandom random(wildChunkSeed);
-
-			// Make sure the coordinate math is done in the new coordinate system.
-			const OriginalInt2 relativeOrigin(
-				((RMDFile::DEPTH - 1) - wildX) * RMDFile::DEPTH,
-				((RMDFile::WIDTH - 1) - wildY) * RMDFile::WIDTH);
-			const NewInt2 dstPoint(
-				relativeOrigin.y + (RMDFile::WIDTH - 1 - x),
-				relativeOrigin.x + (RMDFile::DEPTH - 1 - z));
-
-			// See if the current voxel is a *MENU block and matches the target menu type.
-			const bool matchesTargetType = [this, &dstPoint, menuType]()
-			{
-				const auto &voxelGrid = this->getVoxelGrid();
-				const bool isCity = false; // Wilderness only.
-				const uint16_t voxelID = voxelGrid.getVoxel(dstPoint.x, 1, dstPoint.y);
-				const VoxelDefinition &voxelDef = voxelGrid.getVoxelDef(voxelID);
-				return (voxelDef.dataType == VoxelDataType::Wall) && voxelDef.wall.isMenu() &&
-					(VoxelDefinition::WallData::getMenuType(voxelDef.wall.menuID, isCity) == menuType);
-			}();
-
-			if (matchesTargetType)
-			{
-				// Get the *MENU block's display name.
-				const std::string name = [menuType, &random, &createTavernName, &createTempleName]()
-				{
-					if (menuType == VoxelDefinition::WallData::MenuType::Tavern)
-					{
-						// Tavern.
-						const int m = random.next() % 23;
-						const int n = random.next() % 23;
-						return createTavernName(m, n);
-					}
-					else
-					{
-						// Temple.
-						const int model = random.next() % 3;
-						constexpr std::array<int, 3> ModelVars = { 5, 9, 10 };
-						const int vars = ModelVars.at(model);
-						const int n = random.next() % vars;
-						return createTempleName(model, n);
-					}
-				}();
-
-				this->menuNames.push_back(std::make_pair(dstPoint, std::move(name)));
-			}
-		};
-
-		// Iterate blocks in the chunk in any order. They are order-independent in the wild.
-		for (SNInt x = 0; x < RMDFile::DEPTH; x++)
-		{
-			for (WEInt z = 0; z < RMDFile::WIDTH; z++)
-			{
-				tryGenerateBlockName(x, z);
-			}
-		}
-	};
-
-	// Iterate over each wild chunk.
-	const int wildChunksPerSide = 64;
-	for (int y = 0; y < wildChunksPerSide; y++)
-	{
-		for (int x = 0; x < wildChunksPerSide; x++)
-		{
-			generateNames(x, y, VoxelDefinition::WallData::MenuType::Tavern);
-			generateNames(x, y, VoxelDefinition::WallData::MenuType::Temple);
-		}
-	}
-}
-
 ExteriorLevelData ExteriorLevelData::loadCity(const LocationDefinition &locationDef,
 	const ProvinceDefinition &provinceDef, const MIFFile::Level &level, WeatherType weatherType,
 	int currentDay, int starCount, const std::string &infName, SNInt gridWidth, WEInt gridDepth,
@@ -431,8 +96,8 @@ ExteriorLevelData ExteriorLevelData::loadCity(const LocationDefinition &location
 
 	// Generate building names.
 	const bool isCity = true;
-	levelData.generateBuildingNames(locationDef, provinceDef, random, isCity,
-		gridWidth, gridDepth, miscAssets);
+	levelData.menuNames = CityLevelUtils::generateBuildingNames(locationDef, provinceDef, random,
+		isCity, levelData.getVoxelGrid(), miscAssets);
 
 	// Generate distant sky.
 	levelData.distantSky.init(locationDef, provinceDef, weatherType, currentDay,
@@ -522,7 +187,8 @@ ExteriorLevelData ExteriorLevelData::loadWilderness(const LocationDefinition &lo
 	levelData.readMAP2(tempMap2View, inf);
 
 	// Generate wilderness building names.
-	levelData.generateWildChunkBuildingNames(exeData);
+	levelData.menuNames = WildLevelUtils::generateWildChunkBuildingNames(
+		levelData.getVoxelGrid(), exeData);
 
 	// Generate distant sky.
 	levelData.distantSky.init(locationDef, provinceDef, weatherType, currentDay,
@@ -531,7 +197,7 @@ ExteriorLevelData ExteriorLevelData::loadWilderness(const LocationDefinition &lo
 	return levelData;
 }
 
-const std::vector<std::pair<NewInt2, std::string>> &ExteriorLevelData::getMenuNames() const
+const LevelUtils::MenuNamesList &ExteriorLevelData::getMenuNames() const
 {
 	return this->menuNames;
 }
