+++ OpenTESArena/src/World/WildLevelUtils.cpp
@@ -1,6 +1,9 @@
 #include <algorithm>
 
 #include "CityLevelUtils.h"
+#include "VoxelDataType.h"
+#include "VoxelDefinition.h"
+#include "VoxelGrid.h"
 #include "WildLevelUtils.h"
 #include "../Assets/RMDFile.h"
 #include "../Math/Random.h"
@@ -73,6 +76,130 @@ Buffer2D<WildBlockID> WildLevelUtils::generateWildernessIndices(uint32_t wildSee
 	indices.set(cityX + 1, cityY + 1, 4);
 
 	return indices;
+}
+
+LevelUtils::MenuNamesList WildLevelUtils::generateWildChunkBuildingNames(
+	const VoxelGrid &voxelGrid, const ExeData &exeData)
+{
+	LevelUtils::MenuNamesList menuNames;
+
+	// Lambda for looping through main-floor voxels and generating names for *MENU blocks that
+	// match the given menu type.
+	auto generateNames = [&voxelGrid, &exeData, &menuNames](int wildX, int wildY,
+		VoxelDefinition::WallData::MenuType menuType)
+	{
+		const uint32_t wildChunkSeed = (wildY << 16) + wildX;
+
+		// Don't need hashInSeen() for the wilderness.
+
+		// Lambdas for creating tavern and temple building names.
+		auto createTavernName = [&exeData](int m, int n)
+		{
+			const auto &tavernPrefixes = exeData.cityGen.tavernPrefixes;
+			const auto &tavernSuffixes = exeData.cityGen.tavernSuffixes;
+			return tavernPrefixes.at(m) + ' ' + tavernSuffixes.at(n);
+		};
+
+		auto createTempleName = [&exeData](int model, int n)
+		{
+			const auto &templePrefixes = exeData.cityGen.templePrefixes;
+			const auto &temple1Suffixes = exeData.cityGen.temple1Suffixes;
+			const auto &temple2Suffixes = exeData.cityGen.temple2Suffixes;
+			const auto &temple3Suffixes = exeData.cityGen.temple3Suffixes;
+
+			const std::string &templeSuffix = [&temple1Suffixes, &temple2Suffixes,
+				&temple3Suffixes, model, n]() -> const std::string&
+			{
+				if (model == 0)
+				{
+					return temple1Suffixes.at(n);
+				}
+				else if (model == 1)
+				{
+					return temple2Suffixes.at(n);
+				}
+				else
+				{
+					return temple3Suffixes.at(n);
+				}
+			}();
+
+			// No extra whitespace needed, I think?
+			return templePrefixes.at(model) + templeSuffix;
+		};
+
+		// The lambda called for each main-floor voxel in the area.
+		auto tryGenerateBlockName = [&voxelGrid, &menuNames, wildX, wildY, menuType, wildChunkSeed,
+			&createTavernName, &createTempleName](SNInt x, WEInt z)
+		{
+			ArenaRandom random(wildChunkSeed);
+
+			// Make sure the coordinate math is done in the new coordinate system.
+			const OriginalInt2 relativeOrigin(
+				((RMDFile::DEPTH - 1) - wildX) * RMDFile::DEPTH,
+				((RMDFile::WIDTH - 1) - wildY) * RMDFile::WIDTH);
+			const NewInt2 dstPoint(
+				relativeOrigin.y + (RMDFile::WIDTH - 1 - x),
+				relativeOrigin.x + (RMDFile::DEPTH - 1 - z));
+
+			// See if the current voxel is a *MENU block and matches the target menu type.
+			const bool matchesTargetType = [&voxelGrid, menuType, &dstPoint]()
+			{
+				const bool isCity = false; // Wilderness only.
+				const uint16_t voxelID = voxelGrid.getVoxel(dstPoint.x, 1, dstPoint.y);
+				const VoxelDefinition &voxelDef = voxelGrid.getVoxelDef(voxelID);
+				return (voxelDef.dataType == VoxelDataType::Wall) && voxelDef.wall.isMenu() &&
+					(VoxelDefinition::WallData::getMenuType(voxelDef.wall.menuID, isCity) == menuType);
+			}();
+
+			if (matchesTargetType)
+			{
+				// Get the *MENU block's display name.
+				const std::string name = [menuType, &random, &createTavernName, &createTempleName]()
+				{
+					if (menuType == VoxelDefinition::WallData::MenuType::Tavern)
+					{
+						// Tavern.
+						const int m = random.next() % 23;
+						const int n = random.next() % 23;
+						return createTavernName(m, n);
+					}
+					else
+					{
+						// Temple.
+						const int model = random.next() % 3;
+						constexpr std::array<int, 3> ModelVars = { 5, 9, 10 };
+						const int vars = ModelVars.at(model);
+						const int n = random.next() % vars;
+						return createTempleName(model, n);
+					}
+				}();
+
+				menuNames.push_back(std::make_pair(dstPoint, std::move(name)));
+			}
+		};
+
+		// Iterate blocks in the chunk in any order. They are order-independent in the wild.
+		for (SNInt x = 0; x < RMDFile::DEPTH; x++)
+		{
+			for (WEInt z = 0; z < RMDFile::WIDTH; z++)
+			{
+				tryGenerateBlockName(x, z);
+			}
+		}
+	};
+
+	// Iterate over each wild chunk.
+	for (int y = 0; y < WildLevelUtils::WILD_HEIGHT; y++)
+	{
+		for (int x = 0; x < WildLevelUtils::WILD_WIDTH; x++)
+		{
+			generateNames(x, y, VoxelDefinition::WallData::MenuType::Tavern);
+			generateNames(x, y, VoxelDefinition::WallData::MenuType::Temple);
+		}
+	}
+
+	return menuNames;
 }
 
 void WildLevelUtils::reviseWildernessCity(const LocationDefinition &locationDef,
